- remove color -picker from editor and use Qt's color picker dialog (hi-color we are) - fix color-keys in player edition to use the pre-defined ones, no need to pick it anymore - more changes in file-format - remove destrin-ship scene as this will be handled by movie editor in the future. - move several more hardcoded in-game strings to strings-table

+++ file/format/st_common.h
@@ -194,9 +194,9 @@ struct st_rectangle {
  *
  */
 struct st_color {
-    short r;
-    short g;
-    short b;
+    int r;
+    int g;
+    int b;
 	st_color () {
 		r = -1;
 		g = -1;
